web/app/requirements.txt
amqp==2.6.1
asgiref==3.3.1
astroid==2.5
billiard==3.6.3.0
boto3==1.17.17
botocore==1.20.17
celery==4.4.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
defusedxml==0.7.0rc2
Django==3.1.6
django-storages==1.11.1
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-jsonapi==4.0.0
djangorestframework-jwt
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-yasg==1.20.0
flower==0.9.7
humanize==3.2.0
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.1
isort==5.7.0
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
kombu==4.6.11
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.9
Pillow==8.1.0
prometheus-client==0.8.0
prompt-toolkit==3.0.16
pycparser==2.20
PyJWT
pylint==2.7.0
pylint-django==2.4.2
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-resize-image==1.1.19
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.4
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
typed-ast==1.4.2
uritemplate==3.0.1
urllib3==1.26.3
vine==1.3.0
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.4.0
